+++ src/Soap/DisableEncoder.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Biplane\YandexDirect\Soap;
+
+interface DisableEncoder
+{
+}